+++ spec/unit/generators/templates/rspec_spec.rb
@@ -71,11 +71,6 @@ RSpec.describe Howitzer::BaseGenerator do
             # See https://appium.io/docs/en/about-appium/getting-started/
             # gem 'appium_capybara'
 
-            # Uncomment it if you are going to use 'webkit' driver. QT library should be installed.
-            # See https://github.com/thoughtbot/capybara-webkit/wiki/Installing-Qt-and-compiling-capybara-webkit
-            #
-            # gem 'capybara-webkit'
-
             # Uncomment it if you are going to use 'poltergeist' driver. PhantomJS should be installed.
             # See https://github.com/jnicklas/capybara#poltergeist
             # gem 'poltergeist', git: 'https://github.com/teampoltergeist/poltergeist.git', branch: :master